// src/ng/app/models/paper_attendance_document.ts
export interface PaperAttendanceDocument {
    paper_attendance_docs_id: string;
    dtTraining: string;
    intTrainingCourse: string;
    intUploadedBy: string;
    strOriginalfilename: string;
    id: string;
    compliance_kpis_id: string;
    dtUploaded: string;
    location_name: string;
    is_building: boolean;
    downloadUrl: string;
}